function/requirements.txt
functions-framework==3.*
requests
google-cloud-iam
google-api-python-client
cloudevents
slack-sdk>=3.12,<4
